requirements.txt
streamlit
pandas
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
opencv-python-headless
ultralytics
openai-whisper
